requirements.txt
aiohttp==3.9.5
psycopg[binary]==3.1.19
aiosignal==1.3.1
altgraph==0.17.3
amqp==5.2.0
asgiref==3.7.2
aspose-cells==23.9.0
aspose-words==23.9.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.12.1
beautifulsoup4==4.12.2
billiard==4.2.0
bs4==0.0.1
cachetools==5.3.2
camelot-py==0.9.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==41.0.3
cssselect==1.2.0
cssutils==2.9.0
Django==4.2.4
django-admin-global-search==0.0.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-colorfield==0.9.0
django-cors-headers==4.3.1
django-mathfilters==1.0.0
django-money==3.4.1
django-simple-history==3.4.0
django-sql-utils==0.7.0
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dnspython==2.5.0
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.7
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
eventlet==0.35.1
exceptiongroup==1.2.1
fpdf==1.7.2
frozenlist==1.4.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==5.12.4
html5lib==1.1
htmldocx==0.0.6
httplib2==0.22.0
hurry.filesize==0.9
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
JPype1==1.4.1
keyboard==0.13.5
kombu==5.3.5
Levenshtein==0.21.1
lxml==4.9.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.5
num2words==0.5.12
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pdf2image==1.16.3
pdf2txt==0.7.14
pdfminer.six==20221105
PDFNetPython3==9.5.0.post1
pdfplumber==0.10.3
pefile==2023.2.7
pika==0.13.0
Pillow==10.0.0
pluggy==1.4.0
plyer==2.1.0
premailer==3.10.0
prompt-toolkit==3.0.43
py-moneyed==3.0
PyAMQP==0.1.0.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.18.0
pygame==2.5.0
PyGetWindow==0.0.9
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PyMsgBox==1.0.9
PyMuPDF==1.22.5
pynput==1.7.6
pyparsing==3.1.1
pypdf==3.15.0
PyPDF2==2.12.1
pypdfium2==4.25.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
pytest==8.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2024.1
pywin32-ctypes==0.2.2
rapidfuzz==3.2.0
redis==5.0.1
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
